update code to generate edit/delete buttons dynamically
