+++ src/scripts/login_register.js
@@ -14,7 +14,7 @@ return `
 </div>
 </fieldset>
 `
-}
+}   
 //Builds intial Login Screen//
 function loginUser () {
 return `
